LogLine: Expose via loglistcontext
