+++ public/app/features/logs/components/panel/LogListContext.tsx
@@ -11,6 +11,7 @@ import {
   useState,
 } from 'react';
 
+import { createContext as createAssistantContext, ItemDataType, useAssistant } from '@grafana/assistant';
 import {
   CoreApp,
   DataFrame,
@@ -22,10 +23,11 @@ import {
   shallowCompare,
   store,
 } from '@grafana/data';
-import { config } from '@grafana/runtime';
+import { t } from '@grafana/i18n';
+import { config, getDataSourceSrv } from '@grafana/runtime';
 import { PopoverContent } from '@grafana/ui';
 
-import { DownloadFormat, checkLogsError, checkLogsSampled, downloadLogs as download } from '../../utils';
+import { checkLogsError, checkLogsSampled, downloadLogs as download, DownloadFormat } from '../../utils';
 import { getDisplayedFieldsForLogs } from '../otel/formats';
 
 import { LogLineDetailsMode } from './LogLineDetails';
@@ -99,6 +101,8 @@ export const LogListContext = createContext<LogListContextData>({
   syntaxHighlighting: true,
   toggleDetails: () => {},
   wrapLogMessage: false,
+  isAssistantAvailable: false,
+  openAssistantByLog: () => {},
 });
 
 export const useLogListContextData = (key: keyof LogListContextData) => {
@@ -176,6 +180,8 @@ export interface Props {
   sortOrder: LogsSortOrder;
   syntaxHighlighting?: boolean;
   wrapLogMessage: boolean;
+  isAssistantAvailable: boolean;
+  openAssistantByLog: ((log: LogListModel) => void) | undefined;
 }
 
 export const LogListContextProvider = ({
@@ -236,6 +242,47 @@ export const LogListContextProvider = ({
   const [showDetails, setShowDetails] = useState<LogListModel[]>([]);
   const [detailsWidth, setDetailsWidthState] = useState(getDetailsWidth(containerElement, logOptionsStorageKey));
   const [detailsMode, setDetailsMode] = useState<LogLineDetailsMode>(detailsModeProp ?? 'sidebar');
+  const [isAssistantAvailable, openAssistant] = useAssistant();
+
+  const openAssistantByLog = useCallback(
+    async (log: LogListModel) => {
+      if (!openAssistant) {
+        return;
+      }
+
+      const datasource = await getDataSourceSrv().get(log.datasourceUid);
+      const context = [];
+      if (datasource) {
+        context.push(
+          createAssistantContext(ItemDataType.Datasource, {
+            datasourceUid: datasource.uid,
+            datasourceName: datasource.name,
+            datasourceType: datasource.type,
+          })
+        );
+      }
+      openAssistant({
+        prompt: `${t('logs.log-line-menu.log-line-explainer', 'Explain this log line in a concise way')}:
+
+      \`\`\`
+${log.entry.replaceAll('`', '\\`')}
+      \`\`\`
+      `,
+        context: [
+          ...context,
+          createAssistantContext(ItemDataType.Structured, {
+            title: t('logs.log-line-menu.log-line', 'Log line'),
+            data: {
+              labels: log.labels,
+              value: log.entry,
+              timestamp: log.timestamp,
+            },
+          }),
+        ],
+      });
+    },
+    [openAssistant]
+  );
 
   useEffect(() => {
     if (displayedFields.length > 0 || !config.featureToggles.otelLogsFormatting || !setDisplayedFields) {
@@ -547,6 +594,8 @@ export const LogListContextProvider = ({
         syntaxHighlighting: logListState.syntaxHighlighting,
         toggleDetails,
         wrapLogMessage: logListState.wrapLogMessage,
+        isAssistantAvailable,
+        openAssistantByLog,
       }}
     >
       {children}
